--- client-src/utils/getCurrentScriptSource.js
@@ -1,29 +1,0 @@
-/**
- * @returns {string}
- */
-function getCurrentScriptSource() {
-  // `document.currentScript` is the most accurate way to find the current script,
-  // but is not supported in all browsers.
-  if (document.currentScript) {
-    return document.currentScript.getAttribute("src");
-  }
-
-  // Fallback to getting all scripts running in the document.
-  const scriptElements = document.scripts || [];
-  const scriptElementsWithSrc = Array.prototype.filter.call(
-    scriptElements,
-    (element) => element.getAttribute("src"),
-  );
-
-  if (scriptElementsWithSrc.length > 0) {
-    const currentScript =
-      scriptElementsWithSrc[scriptElementsWithSrc.length - 1];
-
-    return currentScript.getAttribute("src");
-  }
-
-  // Fail as there was no script to use.
-  throw new Error("[webpack-dev-server] Failed to get current script source.");
-}
-
-export default getCurrentScriptSource;
